requirements.txt
<<<<<<< HEAD
am==0.0.1
composer==0.1
dill==0.3.3
kornia==0.6.4
matplotlib==3.5.1
monai==0.8.1
mosaicml==0.5.0
nibabel==3.2.2
numpy==1.21.5
opencv_python==4.5.5.64
pandas==1.4.1
PyYAML==6.0
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.11.1
setuptools==58.0.4
submitit==1.2.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
=======
matplotlib==3.5.1
monai==0.8.1
numpy==1.21.5
PyYAML==6.0
seaborn==0.11.1
torch==1.11.0
>>>>>>> 05210ec13073bcca9b4dbff798fb626d963082dc
